Move class to separate file and remove type hints
